LayoutTests/imported/w3c/web-platform-tests/encrypted-media/drm-expiration.https-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: TypeError: undefined is not an object (evaluating 'supported[ 0 ].content')

Harness Error (FAIL), message = Unhandled rejection: undefined is not an object (evaluating 'supported[ 0 ].content')
